python code refactoring and cleanups
- feat: added argparse
- feat: added several inline paramenters
- feat: added custom settings loader
- feat: added setup.py
