+++ .gitignore
@@ -2,3 +2,8 @@ env
 .env
 venv
 .venv
+*egg-info
+*__pycache__*
+custom_settings/*
+build/*
+dist/*
