+++ update-all-examples.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
-python3 -m demo.simple --replace --no-randomness && \
-python3 -m demo.simple_structured --replace --no-randomness && \
-python3 -m demo.complex_structured --replace --no-randomness
+sd_jwt --replace-examples-in main.md --iat 1516239022 --exp 1516247022 --no-randomness && \
+sd_jwt --replace-examples-in main.md --iat 1516239022 --exp 1516247022 --no-randomness && \
+sd_jwt --replace-examples-in main.md --iat 1516239022 --exp 1516247022  --no-randomness
